MinPlatformPkg/SpiFvbService: Allow dynamic flash info PCDs
The current implementation assumes PCDs containing information about
FV base and size such as those for the variable store FV are
FixedAtBuild. These PCDs and others could be treated as dynamic by
a particular platform so the module is updated to read dynamic (in
addition to FixedAtBuild) PCDs.

In addition, dynamic PCDs are not available in Standalone MM. So
the code also checks if a variable flash information HOB is present
before consuming the flash information from PCDs.

This change focuses on the variable base and size as these are the
primary configuration values adjusted by platforms and directly
consumed by the UEFI varible drivers. Future changes might add
similar support for the fault tolerant PCDs and microcode PCDs.